skymarket/pyproject.toml
[tool.poetry]
name = "skymarket"
version = "0.1.0"
description = ""
authors = ["dev <rvotaisrvota@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest = "^7.2.1"
pytest-factoryboy = "^2.5.1"
drf-spectacular = "^0.25.1"
django-cors-headers = "^3.13.0"
django-phonenumber-field = "^7.0.2"
phonenumbers = "^8.13.4"
djoser = "^2.1.0"
django-rest-framework-redocs = "^0.0.8"
django-filter = "^22.1"
django = "^4.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
